+++ router/router.go
@@ -1,14 +1,11 @@
 package router
 
 import (
-	"encoding/json"
 	"fmt"
 	"math"
 	"math/rand"
 	"sort"
-	"strconv"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/cenkalti/backoff/v4"
@@ -19,11 +16,23 @@ import (
 	"github.com/rudderlabs/rudder-server/jobsdb"
 	"github.com/rudderlabs/rudder-server/rruntime"
 	"github.com/rudderlabs/rudder-server/services/customdestinationmanager"
-	destinationdebugger "github.com/rudderlabs/rudder-server/services/destination-debugger"
 	"github.com/rudderlabs/rudder-server/services/stats"
 	"github.com/rudderlabs/rudder-server/utils/logger"
 	"github.com/rudderlabs/rudder-server/utils/misc"
 )
+
+type WorkerFactory interface {
+	CreateWorkers(rt *HandleT)
+	GetWorker(idx int) WorkerInterface
+	ResetSleep()
+}
+
+type WorkerInterface interface {
+	PostJobOnWorker(job *jobsdb.JobT)
+	WorkerProcess(rt *HandleT)
+	GetFailedJobIDMutex() *sync.RWMutex
+	GetFailedJobIDMap() map[string]int64
+}
 
 //HandleT is the handle to this module.
 type HandleT struct {
@@ -32,7 +41,7 @@ type HandleT struct {
 	jobsDB                   *jobsdb.HandleT
 	netHandle                *NetHandleT
 	destName                 string
-	workers                  []*workerT
+	workerFactory            WorkerFactory
 	perfStats                *misc.PerfStats
 	successCount             uint64
 	failCount                uint64
@@ -60,18 +69,6 @@ type JobParametersT struct {
 	SourceID      string `json:"source_id"`
 	DestinationID string `json:"destination_id"`
 	ReceivedAt    string `json:"received_at"`
-}
-
-// workerT a structure to define a worker for sending events to sinks
-type workerT struct {
-	channel          chan []*jobsdb.JobT // the worker job channel
-	workerID         int                 // identifies the worker
-	failedJobs       int                 // counts the failed jobs of a worker till it gets reset by external channel
-	sleepTime        time.Duration       //the sleep duration for every job of the worker
-	failedJobIDMap   map[string]int64    //user to failed jobId
-	failedJobIDMutex sync.RWMutex        //lock to protect structure above
-	retryForJobMap   map[int64]time.Time //jobID to next retry time map
-	jobRequestQ      chan *jobsdb.JobT
 }
 
 var (
@@ -136,252 +133,6 @@ func (rt *HandleT) trackStuckDelivery() chan struct{} {
 		}
 	})
 	return ch
-}
-
-func (rt *HandleT) workerJobBatcher(worker *workerT) {
-	var workerJobBatchBuffer = make([]*jobsdb.JobT, 0)
-	timeout := time.After(workerRequestBatchTimeout)
-	for {
-		select {
-		case workerJobRequest := <-worker.jobRequestQ:
-			//Append to request buffer
-			workerJobBatchBuffer = append(workerJobBatchBuffer, workerJobRequest)
-			if len(workerJobBatchBuffer) == 100 { //TODO make it configurable
-				worker.channel <- workerJobBatchBuffer
-				workerJobBatchBuffer = nil
-				workerJobBatchBuffer = make([]*jobsdb.JobT, 0)
-			}
-		case <-timeout:
-			timeout = time.After(workerRequestBatchTimeout)
-			if len(workerJobBatchBuffer) > 0 {
-				worker.channel <- workerJobBatchBuffer
-				workerJobBatchBuffer = nil
-				workerJobBatchBuffer = make([]*jobsdb.JobT, 0)
-			}
-		}
-	}
-}
-
-func (rt *HandleT) workerProcess(worker *workerT) {
-	batchTimeStat := stats.NewStat(
-		fmt.Sprintf("router.%s_batch_time", rt.destName), stats.TimerType)
-
-	for {
-		jobs := <-worker.channel
-
-		jobListToSend := make([]*jobsdb.JobT, 0)
-		batchTimeStat.Start()
-
-		for _, job := range jobs {
-			logger.Debugf("[%v Router] :: performing checks to send payload to %s. Payload: ", rt.destName, job.EventPayload)
-
-			userID := job.UserID
-
-			var paramaters JobParametersT
-			err := json.Unmarshal(job.Parameters, &paramaters)
-			if err != nil {
-				logger.Error("Unmarshal of job parameters failed. ", string(job.Parameters))
-			}
-
-			var isPrevFailedUser bool
-			var previousFailedJobID int64
-			if rt.keepOrderOnFailure {
-				//If there is a failed jobID from this user, we cannot pass future jobs
-				worker.failedJobIDMutex.RLock()
-				previousFailedJobID, isPrevFailedUser = worker.failedJobIDMap[userID]
-				worker.failedJobIDMutex.RUnlock()
-
-				// mark job as waiting if prev job from same user has not succeeded yet
-				if isPrevFailedUser {
-					markedAsWaiting := rt.handleJobForPrevFailedUser(job, paramaters, userID, worker, previousFailedJobID)
-					if markedAsWaiting {
-						continue
-					}
-				}
-			}
-
-			// mark job as failed (without incrementing attempts) if same job has failed before and backoff duration not elapsed
-			shouldBackoff := rt.handleBackoff(job, userID, worker)
-			if shouldBackoff {
-				continue
-			}
-
-			// mark job as throttled if either dest level event limit reached or dest user level limit reached
-			hasBeenThrottled := rt.handleThrottle(job, paramaters, userID, worker, isPrevFailedUser)
-			if hasBeenThrottled {
-				continue
-			}
-
-			jobListToSend = append(jobListToSend, job)
-		}
-
-		if len(jobListToSend) == 0 {
-			continue
-		}
-
-		workerVersion := getRouterConfigInt("workerVersion", rt.destName, 0)
-		if workerVersion == 1 {
-		} else {
-			for _, job := range jobs {
-				rt.handleSequential(job, worker)
-			}
-		}
-
-		batchTimeStat.End()
-	}
-}
-
-func (rt *HandleT) handleSequential(job *jobsdb.JobT, worker *workerT) {
-	deliveryTimeStat := stats.NewStat(
-		fmt.Sprintf("router.%s_delivery_time", rt.destName), stats.TimerType)
-	retryAttemptsStat := stats.NewStat(
-		fmt.Sprintf("router.%s_retry_attempts", rt.destName), stats.CountType)
-
-	var respStatusCode, attempts int
-	var respBody string
-	// START: request to destination endpoint
-
-	logger.Debugf("[%v Router] :: trying to send payload. Attempt no. %v of max attempts %v", rt.destName, attempts, ser)
-
-	diagnosisStartTime := time.Now()
-	deliveryTimeStat.Start()
-
-	if job.LastJobStatus.AttemptNum > 0 {
-		retryAttemptsStat.Increment()
-	}
-
-	var paramaters JobParametersT
-	err := json.Unmarshal(job.Parameters, &paramaters)
-	if err != nil {
-		logger.Error("Unmarshal of job parameters failed. ", string(job.Parameters))
-	}
-
-	if rt.customDestinationManager != nil {
-		respStatusCode, _, respBody = rt.customDestinationManager.SendData(job.EventPayload, paramaters.SourceID, paramaters.DestinationID)
-	} else {
-		ch := rt.trackStuckDelivery()
-		respStatusCode, _, respBody = rt.netHandle.sendPost(job.EventPayload)
-		ch <- struct{}{}
-	}
-
-	deliveryTimeStat.End()
-
-	// END: request to destination endpoint
-	rt.handleDestinationResponse(worker, job, diagnosisStartTime, respStatusCode, respBody)
-}
-
-func (rt *HandleT) handleDestinationResponse(worker *workerT, job *jobsdb.JobT, diagnosisStartTime time.Time, respStatusCode int, respBody string) {
-	eventsAbortedStat := stats.NewStat(
-		fmt.Sprintf("router.%s_events_aborted", rt.destName), stats.CountType)
-
-	userID := job.UserID
-
-	_, isPrevFailedUser := worker.failedJobIDMap[userID]
-
-	var paramaters JobParametersT
-	err := json.Unmarshal(job.Parameters, &paramaters)
-	if err != nil {
-		logger.Error("Unmarshal of job parameters failed. ", string(job.Parameters))
-	}
-
-	var reqMetric requestMetric
-	status := jobsdb.JobStatusT{
-		JobID:         job.JobID,
-		ExecTime:      time.Now(),
-		RetryTime:     time.Now(),
-		AttemptNum:    job.LastJobStatus.AttemptNum + 1,
-		ErrorCode:     strconv.Itoa(respStatusCode),
-		ErrorResponse: []byte(`{}`),
-	}
-
-	routerResponseStat := stats.GetRouterStat("router_response_counts", stats.CountType, rt.destName, respStatusCode)
-	routerResponseStat.Increment()
-
-	if isSuccessStatus(respStatusCode) {
-		atomic.AddUint64(&rt.successCount, 1)
-		status.JobState = jobsdb.Succeeded.State
-		reqMetric.RequestSuccess = reqMetric.RequestSuccess + 1
-		reqMetric.RequestCompletedTime = time.Now().Sub(diagnosisStartTime)
-		logger.Debugf("[%v Router] :: sending success status to response", rt.destName)
-		rt.responseQ <- jobResponseT{status: &status, worker: worker, userID: userID}
-	} else {
-		// the job failed
-		logger.Debugf("[%v Router] :: Job failed to send, analyzing...", rt.destName)
-		worker.failedJobs++
-		atomic.AddUint64(&rt.failCount, 1)
-		status.ErrorResponse = []byte(fmt.Sprintf(`{"reason": %v}`, strconv.Quote(respBody)))
-
-		//addToFailedMap is used to decide whethere the jobID has to be added to the failedJobIDMap.
-		//If the job is aborted then there is no point in adding it to the failedJobIDMap.
-		addToFailedMap := true
-
-		status.JobState = jobsdb.Failed.State
-
-		// TODO: 429 should mark this throttled?
-		if respStatusCode >= 500 || respStatusCode == 429 {
-			// TODO: timeElapsed should be ideally from first attempt
-			timeElapsed := time.Now().Sub(job.CreatedAt)
-			if timeElapsed > retryTimeWindow && status.AttemptNum >= maxFailedCountForJob {
-				status.JobState = jobsdb.Aborted.State
-				addToFailedMap = false
-				delete(worker.retryForJobMap, job.JobID)
-			} else {
-				worker.retryForJobMap[job.JobID] = time.Now().Add(durationBeforeNextAttempt(status.AttemptNum))
-			}
-		} else {
-			if status.AttemptNum >= maxFailedCountForJob {
-				status.JobState = jobsdb.Aborted.State
-				addToFailedMap = false
-				eventsAbortedStat.Increment()
-			}
-		}
-
-		if addToFailedMap {
-			//#JobOrder (see other #JobOrder comment)
-			if rt.keepOrderOnFailure && !isPrevFailedUser && userID != "" {
-				logger.Errorf("[%v Router] :: userId %v failed for the first time adding to map", rt.destName, userID)
-				worker.failedJobIDMutex.Lock()
-				worker.failedJobIDMap[userID] = job.JobID
-				worker.failedJobIDMutex.Unlock()
-			}
-		}
-		reqMetric.RequestRetries = reqMetric.RequestRetries + 1
-		reqMetric.RequestCompletedTime = time.Now().Sub(diagnosisStartTime)
-		if status.JobState == jobsdb.Aborted.State {
-			reqMetric.RequestAborted = reqMetric.RequestAborted + 1
-		}
-		logger.Debugf("[%v Router] :: sending failed/aborted state as response", rt.destName)
-		rt.responseQ <- jobResponseT{status: &status, worker: worker, userID: userID}
-	}
-
-	if paramaters.ReceivedAt != "" {
-		if status.JobState == jobsdb.Succeeded.State {
-			receivedTime, err := time.Parse(misc.RFC3339Milli, paramaters.ReceivedAt)
-			if err == nil {
-				eventsDeliveryTimeStat := stats.NewTaggedStat(
-					"event_delivery_time", stats.TimerType, map[string]string{
-						"module":   "router",
-						"destType": rt.destName,
-						"id":       paramaters.DestinationID,
-					})
-
-				eventsDeliveryTimeStat.SendTiming(time.Now().Sub(receivedTime))
-			}
-		}
-	}
-	//Sending destination response to config backend
-	if destinationdebugger.HasUploadEnabled(paramaters.DestinationID) {
-		deliveryStatus := destinationdebugger.DeliveryStatusT{
-			DestinationID: paramaters.DestinationID,
-			SourceID:      paramaters.SourceID,
-			Payload:       job.EventPayload,
-			AttemptNum:    status.AttemptNum,
-			JobState:      status.JobState,
-			ErrorCode:     status.ErrorCode,
-			ErrorResponse: status.ErrorResponse,
-		}
-		destinationdebugger.RecordEventDeliveryStatus(paramaters.DestinationID, &deliveryStatus)
-	}
 }
 
 func (rt *HandleT) handleJobForPrevFailedUser(job *jobsdb.JobT, paramaters JobParametersT, userID string, worker *workerT, previousFailedJobID int64) (markedAsWaiting bool) {
@@ -481,32 +232,7 @@ func (rt *HandleT) trackRequestMetrics(reqMetric requestMetric) {
 	}
 }
 
-func (rt *HandleT) initWorkers() {
-	rt.workers = make([]*workerT, noOfWorkers)
-	for i := 0; i < noOfWorkers; i++ {
-		logger.Info("Worker Started", i)
-		var worker *workerT
-		worker = &workerT{
-			channel:        make(chan []*jobsdb.JobT, noOfJobsPerChannel),
-			failedJobIDMap: make(map[string]int64),
-			retryForJobMap: make(map[int64]time.Time),
-			workerID:       i,
-			failedJobs:     0,
-			sleepTime:      minSleep,
-			jobRequestQ:    make(chan *jobsdb.JobT, noOfJobsPerChannel)}
-		rt.workers[i] = worker
-		rruntime.Go(func() {
-			rt.workerProcess(worker)
-		})
-
-		//workerJobBatchers are used when workerVersion >= 1
-		rruntime.Go(func() {
-			rt.workerJobBatcher(worker)
-		})
-	}
-}
-
-func (rt *HandleT) findWorker(job *jobsdb.JobT) *workerT {
+func (rt *HandleT) findWorker(job *jobsdb.JobT) WorkerInterface {
 
 	userID := job.UserID
 
@@ -517,12 +243,13 @@ func (rt *HandleT) findWorker(job *jobsdb.JobT) *workerT {
 		index = int(math.Abs(float64(misc.GetHash(userID) % noOfWorkers)))
 	}
 
-	worker := rt.workers[index]
+	worker := rt.workerFactory.GetWorker(index)
 
 	//#JobOrder (see other #JobOrder comment)
-	worker.failedJobIDMutex.RLock()
-	defer worker.failedJobIDMutex.RUnlock()
-	blockJobID, found := worker.failedJobIDMap[userID]
+	failedJobIDMutex := worker.GetFailedJobIDMutex()
+	failedJobIDMutex.RLock()
+	defer failedJobIDMutex.RUnlock()
+	blockJobID, found := worker.GetFailedJobIDMap()[userID]
 	if !found {
 		return worker
 	}
@@ -540,9 +267,7 @@ func (rt *HandleT) findWorker(job *jobsdb.JobT) *workerT {
 
 // ResetSleep  this makes the workers reset their sleep
 func (rt *HandleT) ResetSleep() {
-	for _, w := range rt.workers {
-		w.sleepTime = minSleep
-	}
+	rt.workerFactory.ResetSleep()
 }
 
 //Enable enables a router :)
@@ -723,12 +448,13 @@ func (rt *HandleT) generatorLoop() {
 		//#JobOrder (See comment marked #JobOrder
 		rt.toClearFailJobIDMutex.Lock()
 		for idx := range rt.toClearFailJobIDMap {
-			wrk := rt.workers[idx]
-			wrk.failedJobIDMutex.Lock()
+			wrk := rt.workerFactory.GetWorker(idx)
+			failedJobIDMutex := wrk.GetFailedJobIDMutex()
+			failedJobIDMutex.Lock()
 			for _, userID := range rt.toClearFailJobIDMap[idx] {
-				delete(wrk.failedJobIDMap, userID)
+				delete(wrk.GetFailedJobIDMap(), userID)
 			}
-			wrk.failedJobIDMutex.Unlock()
+			failedJobIDMutex.Unlock()
 		}
 		rt.toClearFailJobIDMap = make(map[int][]string)
 		rt.toClearFailJobIDMutex.Unlock()
@@ -764,7 +490,7 @@ func (rt *HandleT) generatorLoop() {
 
 		//List of jobs wich can be processed mapped per channel
 		type workerJobT struct {
-			worker *workerT
+			worker WorkerInterface
 			job    *jobsdb.JobT
 		}
 
@@ -794,12 +520,7 @@ func (rt *HandleT) generatorLoop() {
 
 		//Send the jobs to the jobQ
 		for _, wrkJob := range toProcess {
-			workerVersion := getRouterConfigInt("workerVersion", rt.destName, 0)
-			if workerVersion == 1 {
-				wrkJob.worker.jobRequestQ <- wrkJob.job
-			} else {
-				wrkJob.worker.channel <- []*jobsdb.JobT{wrkJob.job}
-			}
+			wrkJob.worker.PostJobOnWorker(wrkJob.job)
 		}
 
 		countStat.Count(len(combinedList))
@@ -847,7 +568,14 @@ func (rt *HandleT) Setup(jobsDB *jobsdb.HandleT, destName string) {
 	throttler.SetUp(rt.destName)
 	rt.throttler = &throttler
 
-	rt.initWorkers()
+	workerVersion := getRouterConfigString("workerVersion", rt.destName, "solo")
+	if workerVersion == "batch" {
+		rt.workerFactory = &BatchWorkerFactory{}
+	} else {
+		rt.workerFactory = &SoloWorkerFactory{}
+	}
+	rt.workerFactory.CreateWorkers(rt)
+
 	rruntime.Go(func() {
 		rt.collectMetrics()
 	})
